fix: merge conflict with roomsummarytopic, and move to vm right_panel
