+++ src/components/viewmodels/right_panel/RoomSummaryCardViewModel.tsx
