+++ test/unit-tests/components/viewmodels/right_panel/RoomSummaryCardViewModel-test.tsx
@@ -8,7 +8,7 @@
 import { act, renderHook, waitFor } from "jest-matrix-react";
 import { type MatrixClient, type Room } from "matrix-js-sdk/src/matrix";
 
-import { useRoomSummaryCardViewModel } from "../../../../../src/components/viewmodels/rooms/RoomSummaryCardViewModel";
+import { useRoomSummaryCardViewModel } from "../../../../../src/components/viewmodels/right_panel/RoomSummaryCardViewModel";
 import { mkStubRoom, stubClient, withClientContextRenderOptions } from "../../../../test-utils";
 import defaultDispatcher from "../../../../../src/dispatcher/dispatcher";
 import RoomListStore from "../../../../../src/stores/room-list/RoomListStore";
